back/requirements.txt
fastapi
uvicorn[standard]
requests
pytest
pytest-cov
python-dotenv